// graphql/generated.ts
export enum GalleryGridType {
  hero = 'hero',
  normal = 'normal',
  simple = 'simple'
}